commit conftest.py and fixed modules test_processing.py and test_widget.py
