+++ tests/conftest.py
@@ -1,0 +1,36 @@
+import pytest
+
+
+@pytest.fixture
+def sample_card_and_account_data() -> list[str]:
+    """Создаем фикстуру для тестирования"""
+    return [
+        "Maestro 1596837868705199",
+        "Счет 64686473678894779589",
+        "MasterCard 7158300734726758",
+        "Счет 35383033474447895560",
+        "Visa Classic 6831982476737658",
+        "Visa Platinum 8990922113665229",
+        "Visa Gold 5999414228426353",
+        "Счет 73654108430135874305",
+    ]
+
+
+@pytest.fixture
+def sample_transaction_data() -> list[dict]:
+    """Создаем фикстуру для тестирования"""
+    return [
+        {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
+        {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
+        {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
+        {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"},
+    ]
+
+
+@pytest.fixture
+def similar_date_transactions() -> list[dict]:
+    """Создаем фикстуру для тестирования по одинаковым датам"""
+    return [
+        {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
+        {"id": 939719570, "state": "EXECUTED", "date": "2019-07-03T18:36:29.512364"},
+    ]
