+++ tests/test_widget.py
@@ -3,21 +3,6 @@ import pytest
 from src.widget import get_date, mask_account_card
 
 
-@pytest.fixture
-def sample_card_and_account_data() -> list[str]:
-    """Создаем фикстуру для тестирования"""
-    return [
-        "Maestro 1596837868705199",
-        "Счет 64686473678894779589",
-        "MasterCard 7158300734726758",
-        "Счет 35383033474447895560",
-        "Visa Classic 6831982476737658",
-        "Visa Platinum 8990922113665229",
-        "Visa Gold 5999414228426353",
-        "Счет 73654108430135874305",
-    ]
-
-
 @pytest.mark.parametrize(
     "sample_card_and_account_data, expected",
     [
